requirements.txt
langchain==0.0.272
openai==0.27.9
PyPDF2==3.0.1
python-dotenv==1.0.0
streamlit==1.25.0
pandas==2.0.3
faiss-cpu==1.7.4
tiktoken==0.4.0